feat: add favorite list API and integrate with MyPage
