+++ backend/users/serializers.py
@@ -1,6 +1,5 @@
-# backend/users/serializers.py
 from rest_framework import serializers
-from .models import User
+from .models import User   # カスタムユーザーモデルを直接import
 
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
